(maint) Extract Puppet Forge function
Extract all puppet forge api calls to a single file. This makes sure we call the forge with the correct headers for all hover providers and provide a consistent interface to the forge.

This makes a cosmetic change to how the metadata.json hover provider shows hover information by making it the same as the pupeptfile hover provider.
